+++ src/shapes/Group.spec.ts
@@ -1,4 +1,5 @@
 import { FixedLayout } from '../LayoutManager';
+import { Canvas } from '../canvas/Canvas';
 import { Group } from './Group';
 import { FabricObject } from './Object/FabricObject';
 
@@ -32,5 +33,41 @@ describe('Group', () => {
       target: group,
       type: 'imperative',
     });
+  });
+  test('adding and removing an object', () => {
+    const object = new FabricObject();
+    const group = new Group([object]);
+    const group2 = new Group();
+    const canvas = new Canvas();
+
+    const eventsSpy = jest.spyOn(object, 'fire');
+    const removeSpy = jest.spyOn(group, 'remove');
+    const exitSpy = jest.spyOn(group, 'exitGroup');
+    const enterSpy = jest.spyOn(group2, 'enterGroup');
+
+    expect(object.group).toBe(group);
+    expect(object.parent).toBe(group);
+    expect(object.canvas).toBeUndefined();
+
+    canvas.add(group, group2);
+    expect(object.canvas).toBe(canvas);
+
+    group2.add(object);
+    expect(object.group).toBe(group2);
+    expect(object.parent).toBe(group2);
+    expect(object.canvas).toBe(canvas);
+    expect(removeSpy).toBeCalledWith(object);
+    expect(exitSpy).toBeCalledWith(object, undefined);
+    expect(enterSpy).toBeCalledWith(object, true);
+    expect(eventsSpy).toHaveBeenNthCalledWith(1, 'removed', { target: group });
+    expect(eventsSpy).toHaveBeenNthCalledWith(2, 'added', { target: group2 });
+
+    group2.remove(object);
+    expect(eventsSpy).toHaveBeenNthCalledWith(3, 'removed', { target: group2 });
+    expect(object.group).toBeUndefined();
+    expect(object.parent).toBeUndefined();
+    expect(object.canvas).toBeUndefined();
+
+    expect(eventsSpy).toBeCalledTimes(3);
   });
 });